requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
boto3==1.26.14
botocore==1.29.14
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
contourpy==1.0.6
cryptography==39.0.0
cycler==0.11.0
diffusers @ git+https://github.com/huggingface/diffusers.git@182eb959e5efc8c77fa31394ca55376331c0ed25
dnspython==2.2.1
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.87.0
ffmpeg-python==0.2.0
ffmpy==0.3.0
filelock==3.8.0
filetype==1.2.0
Flask==2.2.2
Flask-Cors==3.0.10
flatbuffers==22.12.6
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==6.1.1
future==0.18.3
google-auth==2.14.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.2
huggingface-hub==0.11.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.23.0
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==1.0.3
llvmlite==0.39.1
loguru==0.6.0
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
more-itertools==9.0.0
mpmath==1.2.1
multidict==6.0.4
networkx==2.8.8
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.2.3
onnxruntime==1.13.1
openai-whisper==20230124
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
orjson==3.8.3
packaging==22.0
pandas==1.5.2
pandas-stubs==1.5.2.221213
Pillow==9.3.0
pooch==1.6.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.2
pydub==0.25.1
PyJWT==2.6.0
PyMatting==1.1.8
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.5
pytube==12.1.2
pytz==2022.6
pywatchman==1.4.1
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.10.31
rembg==2.0.30
requests==2.28.1
requests-oauthlib==1.3.1
revChatGPT==0.0.14
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
sacremoses==0.0.53
scikit-image==0.19.3
scipy==1.9.3
semantic-version==2.10.0
sentencepiece==0.1.97
setuptools-rust==1.5.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.21.0
sympy==1.11.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2022.10.10
tokenizers==0.13.2
toolz==0.12.0
torch==1.12.1
torchvision==0.14.0
tqdm==4.64.1
transformers==4.26.0
twilio==7.15.4
types-pytz==2022.7.0.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
urllib3==1.26.13
uvicorn==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
